third_party/blink/web_tests/external/wpt/css/css-typed-om/the-stylepropertymap/properties/scroll-padding-expected.txt

This is a testharness.js-based test.
Found 16 FAIL, 0 TIMEOUT, 0 NOTRUN.
[FAIL] Can set 'scroll-padding-top' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-top' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-left' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-left' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-right' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-right' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-bottom' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-bottom' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-inline-start' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-inline-start' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-block-start' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-block-start' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-inline-end' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-inline-end' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-block-end' to a percent: -3.14%
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
[FAIL] Can set 'scroll-padding-block-end' to a length: -3.14em
  assert_equals: expected "CSSMathSum" but got "CSSUnitValue"
Harness: the test ran to completion.
